refactor(global): update by pr comments
