+++ packages/client/src/pages/ErrorPage/ErrorPage.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 
 import { screen } from '@testing-library/react';
 
-import { renderWithRouter, waitUntilLoaderToBeRemoved } from '../../utils/testing-utils';
+import { renderWithRouter, waitUntilLoaderToBeRemoved } from '../../utils/testingUtils';
 
 describe('Error Page', () => {
   test('it renders at a wrong path', async () => {
