+++ packages/client/src/pages/Forum/Forum.tsx
@@ -11,16 +11,7 @@ export const Forum: FC<ForumProps> = () => {
     <section className="forum__wrapper">
       <h1 className="forum__title">Форум</h1>
       <table border={1} className="forum" data-testid="forum">
-        <thead className="forum__row-header">
-          <tr>
-            <th className="forum__cell-header">Разделы</th>
-            <th className="forum__cell-header">Темы</th>
-            <th className="forum__cell-header">Ответы</th>
-          </tr>
-        </thead>
-        <tbody className="forum__body">
-          <ForumSectionList sectionList={sectionList} />
-        </tbody>
+        <ForumSectionList sectionList={sectionList} />
       </table>
     </section>
   );
