+++ packages/client/src/pages/Forum/ForumSection/ForumSection.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 
 import { screen } from '@testing-library/react';
 
-import { renderWithRouter } from '../../../utils/testing-utils';
+import { renderWithRouter } from '../../../utils/testingUtils';
 import { ForumSection } from './ForumSection';
 
 describe('ForumSection', () => {
