+++ packages/client/src/pages/Forum/ForumSectionList/ForumSectionList.tsx
@@ -7,15 +7,24 @@ import { ForumSectionListProps } from './typings';
 export const ForumSectionList: FC<ForumSectionListProps> = ({ sectionList }) => {
   return (
     <>
-      {sectionList.map(item => (
-        <tr key={item.id} data-testid={item.id} className="forum__item">
-          <td className="forum__first-column">
-            <Link to={`${Paths.Section}/${item.id}`}>{item.name}</Link>
-          </td>
-          <td>{item.topicCount}</td>
-          <td>{item.messages}</td>
+      <thead className="forum__row-header">
+        <tr>
+          <th className="forum__cell-header">Разделы</th>
+          <th className="forum__cell-header">Темы</th>
+          <th className="forum__cell-header">Ответы</th>
         </tr>
-      ))}
+      </thead>
+      <tbody className="forum__body">
+        {sectionList.map(item => (
+          <tr key={item.id} data-testid={item.id} className="forum__item">
+            <td className="forum__first-column">
+              <Link to={`${Paths.Section}/${item.id}`}>{item.name}</Link>
+            </td>
+            <td>{item.topicCount}</td>
+            <td>{item.messages}</td>
+          </tr>
+        ))}
+      </tbody>
     </>
   );
 };
