+++ packages/client/src/pages/Forum/ForumTopic/ForumMessage/typings.ts
@@ -1,7 +1,10 @@
-export type ForumMessageProps = {
+type ForumMessage = {
   id: number;
   content: string;
   date: string;
   userId: number;
   username: string;
+};
+export type ForumMessageProps = {
+  message: ForumMessage;
 };
