+++ packages/client/src/pages/Forum/ForumTopic/ForumTopic.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom';
 
 import { screen } from '@testing-library/react';
 
-import { renderWithRouter } from '../../../utils/testing-utils';
+import { renderWithRouter } from '../../../utils/testingUtils';
 import { ForumTopic } from './ForumTopic';
 
 describe('ForumTopic', () => {
