--- packages/client/src/pages/Forum/ForumTopic/typings.ts
@@ -1,6 +1,0 @@
-export type ForumTopicRowProps = {
-  name?: string;
-};
-export type MessageForm = {
-  message: string;
-};
