+++ packages/client/src/utils/HTTP/httpClient.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosError } from 'axios';
 
 import { APIError } from '../../api/typings';
-import { apiHasError } from '../api-utils';
+import { apiHasError } from '../apiUtils';
 
 // Check if error response body contains reason of error
 axios.interceptors.response.use(
